+++ EPlast/EPlast.WebApi/Controllers/AuthController.cs
@@ -13,12 +13,6 @@ namespace EPlast.WebApi.Controllers
     [ApiController]
     public class AuthController : ControllerBase
     {
-        private readonly IAuthEmailService _authEmailServices;
-        private readonly IAuthService _authService;
-        private readonly IHomeService _homeService;
-        private readonly IResources _resources;
-        private readonly IUserDatesService _userDatesService;
-
         public AuthController(
             IAuthService authService,
             IUserDatesService userDatesService,
@@ -75,14 +69,6 @@ namespace EPlast.WebApi.Controllers
             {
                 return Ok(_resources.ResourceForErrors["ConfirmedEmailNotAllowed"]);
             }
-        }
-
-        [HttpGet("logout")]
-        [Authorize(AuthenticationSchemes = "Bearer")]
-        public IActionResult Logout()
-        {
-            _authService.SignOutAsync();
-            return Ok();
         }
 
         /// <summary>
@@ -164,5 +150,11 @@ namespace EPlast.WebApi.Controllers
 
             return Ok(_resources.ResourceForErrors["Feedback-Sended"]);
         }
+
+        private readonly IAuthEmailService _authEmailServices;
+        private readonly IAuthService _authService;
+        private readonly IHomeService _homeService;
+        private readonly IResources _resources;
+        private readonly IUserDatesService _userDatesService;
     }
 }
