+++ EPlast/EPlast.XUnitTest/Services/Authentication/AccountServiceTests.cs
@@ -77,20 +77,52 @@ namespace EPlast.XUnitTest.Services
                .Setup(s => s.Map<UserDTO, User>(It.IsAny<UserDTO>()))
                .Returns(GetTestUserWithEmailsSendedTime());
             Mock<IUrlHelperFactory> mockUrlHelperFactory = new Mock<IUrlHelperFactory>();
-            Mock<IActionContextAccessor> mockActionContextAccessor = new Mock<IActionContextAccessor>(); 
+            Mock<IActionContextAccessor> mockActionContextAccessor = new Mock<IActionContextAccessor>();
             Mock<IHttpContextAccessor> mockHttpContextAccessor = new Mock<IHttpContextAccessor>();
-            
 
-            AuthService AuthService = new AuthService(mockUserManager.Object, 
-                mockSignInManager.Object, 
-                mockEmailConfirmation.Object, 
-                mockMapper.Object, 
-                mockRepositoryWrapper.Object,
-                mockUrlHelperFactory.Object,
-                mockActionContextAccessor.Object,
-                mockHttpContextAccessor.Object);
-                
+            AuthService AuthService = new AuthService(
+                mockUserManager.Object,
+                mockSignInManager.Object,
+                mockEmailConfirmation.Object,
+                mockMapper.Object,
+                mockRepositoryWrapper.Object);
+
             return (mockSignInManager, mockUserManager, mockEmailConfirmation, AuthService);
+        }
+
+        public (
+            Mock<IEmailConfirmation>,
+            Mock<IAuthService>,
+            Mock<UserManager<User>>,
+            Mock<IUrlHelperFactory>,
+            Mock<IActionContextAccessor>,
+            Mock<IHttpContextAccessor>,
+            AuthEmailService
+            ) CreateAuthEmailService()
+        {
+            Mock<IEmailConfirmation> mockEmailConfirmatioService = new Mock<IEmailConfirmation>();
+            Mock<IAuthService> mockAuthSerive = new Mock<IAuthService>();
+            var store = new Mock<IUserStore<User>>();
+            Mock<UserManager<User>> mockUserManager = new Mock<UserManager<User>>(store.Object, null, null, null, null, null, null, null, null);
+            Mock<IUrlHelperFactory> mockUrlFactory = new Mock<IUrlHelperFactory>();
+            Mock<IActionContextAccessor> mockActioAccessor = new Mock<IActionContextAccessor>();
+            Mock<IHttpContextAccessor> mockHttpContextAccessor = new Mock<IHttpContextAccessor>();
+            AuthEmailService authEmailService = new AuthEmailService(
+                mockEmailConfirmatioService.Object,
+                mockAuthSerive.Object,
+                mockUserManager.Object,
+                mockUrlFactory.Object,
+                mockActioAccessor.Object,
+                mockHttpContextAccessor.Object);
+            return (
+                mockEmailConfirmatioService,
+                mockAuthSerive,
+                mockUserManager,
+                mockUrlFactory,
+                mockActioAccessor,
+                mockHttpContextAccessor,
+                authEmailService
+                );
         }
 
         [Fact]
@@ -135,7 +167,14 @@ namespace EPlast.XUnitTest.Services
         public async Task TestConfirmEmail()
         {
             //Arrange
-           var (mockSignInManager, mockUserManager, mockEmailConfirmation, AuthService) = CreateAuthService();
+            var (
+                _,
+                _,
+                mockUserManager,
+                _,
+                _,
+                _,
+                authEmailService) = CreateAuthEmailService();
             mockUserManager
                 .Setup(s => s.FindByIdAsync(It.IsAny<string>()))
                 .ReturnsAsync(GetTestUserWithAllFields());
@@ -145,7 +184,7 @@ namespace EPlast.XUnitTest.Services
               .Returns(Task.FromResult(IdentityResult.Success));
 
             //Act
-            var result = await AuthService.ConfirmEmailAsync(GetTestCode(), GetTestCode());
+            var result = await authEmailService.ConfirmEmailAsync(GetTestCode(), GetTestCode());
 
             //Assert
             var identityResult = Assert.IsType<IdentityResult>(result);
@@ -224,7 +263,7 @@ namespace EPlast.XUnitTest.Services
             Assert.NotNull(authResult);
         }
 
-        [Fact]                                               
+        [Fact]
         public async Task TestIsEmailConfirmedAsync()
         {
             //Arrange
@@ -450,7 +489,7 @@ namespace EPlast.XUnitTest.Services
 
         private AuthenticationProperties GetTestAuthenticationProperties()
         {
-           Dictionary<string, string> authenticationDictionary = new Dictionary<string, string>(3);
+            Dictionary<string, string> authenticationDictionary = new Dictionary<string, string>(3);
             authenticationDictionary.Add("First", "Google");
             authenticationDictionary.Add("Second", "Facebook");
             authenticationDictionary.Add("Third", "Amazon");
@@ -462,6 +501,7 @@ namespace EPlast.XUnitTest.Services
         {
             return new string("andriishainoha@gmail.com");
         }
+
         private string GetTestProvider()
         {
             return new string("fakeProvider");
@@ -476,7 +516,7 @@ namespace EPlast.XUnitTest.Services
 
         private string GetTestCodeForResetPasswordAndConfirmEmail()
         {
-           return new string("500");
+            return new string("500");
         }
 
         private string GetTestIdForUser()
