+++ EPlast/EPlast.sln
@@ -28,6 +28,7 @@ Global
 		{3DBC3BB8-D220-4B4A-81F5-4646F6802007}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3DBC3BB8-D220-4B4A-81F5-4646F6802007}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A1BB5905-A6A7-48AF-A4DE-B5ED52282BA5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A1BB5905-A6A7-48AF-A4DE-B5ED52282BA5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A1BB5905-A6A7-48AF-A4DE-B5ED52282BA5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A1BB5905-A6A7-48AF-A4DE-B5ED52282BA5}.Release|Any CPU.Build.0 = Release|Any CPU
 		{76067960-4AFB-4DC3-B565-552FCF519AAD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
@@ -47,6 +48,7 @@ Global
 		{AB9A904C-A828-4D59-88FA-9D2730A43935}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AB9A904C-A828-4D59-88FA-9D2730A43935}.Release|Any CPU.Build.0 = Release|Any CPU
 		{60967799-6640-4067-B320-E5659FE29A5A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{60967799-6640-4067-B320-E5659FE29A5A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{60967799-6640-4067-B320-E5659FE29A5A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{60967799-6640-4067-B320-E5659FE29A5A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
